Only shuffle block permutation once per block
It is not necessary to re-shuffle the block permutation for every round for each block, and the code is alot simpler if we only shuffle it once at the beginning of each block.  That way we dont need to have an entropy buffer for the shuffling for each round when we decrypt.